Split FfmpegVideoRecorder.cs into focused partial files (#9464)
Extract VideoSegment into its own file and split the recorder across partial files for process management and Windows window-capture, keeping each file well under 300 lines. No behavioral or API changes.
