+++ src/Platform/Microsoft.Testing.Extensions.VideoRecorder/FfmpegVideoRecorder.cs
@@ -2,37 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Microsoft.Testing.Extensions.VideoRecorder.Resources;
-using Microsoft.Testing.Platform;
 using Microsoft.Testing.Platform.Helpers;
 
-#if !NETCOREAPP
-using Polyfills;
-#endif
-
 namespace Microsoft.Testing.Extensions.VideoRecorder;
-
-/// <summary>
-/// A single finalized recording segment and the time range (in seconds, relative to the start of
-/// the recording) that it covers.
-/// </summary>
-internal readonly struct VideoSegment
-{
-    public VideoSegment(string path, double startSeconds, double endSeconds)
-    {
-        Path = path;
-        StartSeconds = startSeconds;
-        EndSeconds = endSeconds;
-    }
-
-    public string Path { get; }
-
-    public double StartSeconds { get; }
-
-    public double EndSeconds { get; }
-
-    public bool Overlaps(double fromSeconds, double toSeconds)
-        => EndSeconds > fromSeconds && StartSeconds < toSeconds;
-}
 
 /// <summary>
 /// Captures the screen by driving a single long-lived ffmpeg process that writes the recording as a
@@ -50,7 +22,7 @@ internal readonly struct VideoSegment
 [UnsupportedOSPlatform("ios")]
 [UnsupportedOSPlatform("tvos")]
 [UnsupportedOSPlatform("wasi")]
-internal sealed class FfmpegVideoRecorder
+internal sealed partial class FfmpegVideoRecorder
 {
     private static readonly TimeSpan StopTimeout = TimeSpan.FromSeconds(15);
 
@@ -373,351 +345,5 @@ internal sealed class FfmpegVideoRecorder
                 || line.IndexOf("Failed", StringComparison.OrdinalIgnoreCase) >= 0);
 
         return errors.Length > 0 ? string.Join(" | ", errors) : lines[lines.Length - 1];
-    }
-
-    private async Task<bool> RunFfmpegAsync(string arguments, CancellationToken cancellationToken)
-    {
-        var startInfo = new ProcessStartInfo(FfmpegPath!, arguments)
-        {
-            RedirectStandardOutput = true,
-            RedirectStandardError = true,
-            UseShellExecute = false,
-            CreateNoWindow = true,
-        };
-
-        _log?.Invoke($"Running ffmpeg: \"{FfmpegPath}\" {arguments}");
-
-        using var process = new Process { StartInfo = startInfo };
-        process.OutputDataReceived += OnFfmpegOutput;
-        process.ErrorDataReceived += OnFfmpegOutput;
-        try
-        {
-            process.Start();
-            process.BeginOutputReadLine();
-            process.BeginErrorReadLine();
-            if (!await WaitForExitAsync(process, StopTimeout, cancellationToken).ConfigureAwait(false))
-            {
-                // ffmpeg didn't finish within the timeout (or the run was cancelled). Disposing the
-                // Process below does not terminate the underlying OS process, so kill it explicitly
-                // to avoid leaving an orphaned ffmpeg behind on slow machines or long sessions.
-                TryKill(process);
-                return false;
-            }
-
-            return process.HasExited && process.ExitCode == 0;
-        }
-        catch (Exception ex)
-        {
-            _log?.Invoke($"ffmpeg invocation failed: {ex.Message}");
-            return false;
-        }
-        finally
-        {
-            process.OutputDataReceived -= OnFfmpegOutput;
-            process.ErrorDataReceived -= OnFfmpegOutput;
-        }
-    }
-
-#if NETCOREAPP
-    private void ClearRecentFfmpegOutput() => _recentFfmpegOutput.Clear();
-#else
-    private void ClearRecentFfmpegOutput()
-    {
-        while (_recentFfmpegOutput.TryDequeue(out _))
-        {
-        }
-    }
-#endif
-
-    private void DetachAndDispose(Process process)
-    {
-        process.OutputDataReceived -= OnFfmpegOutput;
-        process.ErrorDataReceived -= OnFfmpegOutput;
-        process.Dispose();
-    }
-
-    private void OnFfmpegOutput(object sender, DataReceivedEventArgs e)
-    {
-        if (e.Data is null)
-        {
-            return;
-        }
-
-        _log?.Invoke($"[ffmpeg] {e.Data}");
-
-        // Keep a small tail of ffmpeg output so capture failures can be reported with context.
-        _recentFfmpegOutput.Enqueue(e.Data);
-        while (_recentFfmpegOutput.Count > 8 && _recentFfmpegOutput.TryDequeue(out _))
-        {
-        }
-    }
-
-    private string BuildSegmentArguments(string segmentDirectory, string segmentListPath)
-    {
-        int segmentSeconds = Math.Max(1, _options.SegmentLengthSeconds);
-        string input = _options.InputArgumentsOverride ?? BuildDefaultInput();
-        string encoder = _options.Format == VideoRecorderFormat.WebMVp9
-            ? "-c:v libvpx-vp9 -b:v 0 -crf 32 -deadline realtime -cpu-used 5"
-            : "-c:v libx264 -preset veryfast -crf 28";
-        string segmentFormat = _options.Format == VideoRecorderFormat.WebMVp9 ? "webm" : "mp4";
-
-        string extra = RoslynString.IsNullOrWhiteSpace(_options.ExtraRecorderArguments)
-            ? string.Empty
-            : $"{_options.ExtraRecorderArguments!.Trim()} ";
-
-        string segmentPattern = Path.Combine(segmentDirectory, $"seg_%05d.{SegmentExtension}");
-
-        // gdigrab (and some grabbers) can capture an odd width/height, but yuv420p requires both to
-        // be even. Crop to the nearest even dimensions so the encoder's format conversion can't fail
-        // with "width/height not divisible by 2". Skipped when the caller fully overrides the input.
-        string evenCrop = _options.InputArgumentsOverride is null
-            ? "-vf \"crop=trunc(iw/2)*2:trunc(ih/2)*2\" "
-            : string.Empty;
-
-        // Force a keyframe at every segment boundary so each segment starts on an IDR frame; that
-        // makes the segments independently decodable and lets us concatenate them later with
-        // stream copy (-c copy) instead of a slow, lossy re-encode.
-        return $"-y {input} {evenCrop}{encoder} -pix_fmt yuv420p -force_key_frames \"expr:gte(t,n_forced*{segmentSeconds})\" "
-            + $"-f segment -segment_time {segmentSeconds} -segment_format {segmentFormat} -reset_timestamps 1 "
-            + $"-segment_list \"{segmentListPath}\" -segment_list_type csv {extra}\"{segmentPattern}\"";
-    }
-
-    private string BuildDefaultInput()
-    {
-        int fps = _options.FrameRate;
-
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-        {
-            if (_options.Source == VideoCaptureSource.Window)
-            {
-                if (TryGetCurrentProcessWindowRegion(out int x, out int y, out int width, out int height))
-                {
-                    return $"-f gdigrab -framerate {fps} -offset_x {x} -offset_y {y} -video_size {width}x{height} -i desktop";
-                }
-
-                _log?.Invoke(VideoRecorderResources.WindowCaptureFallback);
-            }
-
-            return $"-f gdigrab -framerate {fps} -i desktop";
-        }
-
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-        {
-            // The avfoundation screen device index can vary; override via InputArgumentsOverride if needed.
-            return $"-f avfoundation -capture_cursor 1 -framerate {fps} -i \"Capture screen 0\"";
-        }
-
-        // Assume Linux/X11.
-        string display = Environment.GetEnvironmentVariable("DISPLAY") is { Length: > 0 } d ? d : ":0.0";
-        return $"-f x11grab -framerate {fps} -video_size {_options.X11CaptureSize} -i {display}";
-    }
-
-    // Resolves the screen rectangle of the window to capture so gdigrab can record just that
-    // region. Candidates are tried in order: the process main window (a GUI app under test owns
-    // it), then the foreground window (the terminal you launched from — this is what makes
-    // Windows Terminal work, since its window isn't owned by the test process), then the console
-    // window (classic conhost). Returns false when none is a usable visible window, in which case
-    // the caller falls back to full-screen capture.
-    private static bool TryGetCurrentProcessWindowRegion(out int x, out int y, out int width, out int height)
-    {
-        x = y = width = height = 0;
-
-#if NETCOREAPP
-        if (!OperatingSystem.IsWindows())
-        {
-            return false;
-        }
-#endif
-
-        // gdigrab captures physical pixels, so we must read window rectangles in physical
-        // coordinates too. Make this thread Per-Monitor-V2 DPI aware while querying; otherwise a
-        // DPI-unaware process gets virtualized (logical) coordinates and the region is wrong on
-        // any scaled display. Restore the previous context afterwards.
-        IntPtr previousDpiContext = TrySetPerMonitorDpiAwareThread();
-        try
-        {
-            int screenWidth = NativeMethods.GetSystemMetrics(NativeMethods.SM_CXSCREEN);
-            int screenHeight = NativeMethods.GetSystemMetrics(NativeMethods.SM_CYSCREEN);
-
-            foreach (IntPtr handle in EnumerateCandidateWindows())
-            {
-                if (handle == IntPtr.Zero
-                    || !NativeMethods.IsWindowVisible(handle)
-                    || !NativeMethods.GetWindowRect(handle, out NativeMethods.RECT rect))
-                {
-                    continue;
-                }
-
-                int left = Math.Max(0, rect.Left);
-                int top = Math.Max(0, rect.Top);
-                int right = screenWidth > 0 ? Math.Min(rect.Right, screenWidth) : rect.Right;
-                int bottom = screenHeight > 0 ? Math.Min(rect.Bottom, screenHeight) : rect.Bottom;
-
-                // gdigrab requires even dimensions for the yuv420p pixel format.
-                int candidateWidth = (right - left) & ~1;
-                int candidateHeight = (bottom - top) & ~1;
-                if (candidateWidth <= 0 || candidateHeight <= 0)
-                {
-                    continue;
-                }
-
-                x = left;
-                y = top;
-                width = candidateWidth;
-                height = candidateHeight;
-                return true;
-            }
-        }
-        finally
-        {
-            if (previousDpiContext != IntPtr.Zero)
-            {
-                try
-                {
-                    NativeMethods.SetThreadDpiAwarenessContext(previousDpiContext);
-                }
-                catch (Exception)
-                {
-                    // Best effort; the override is per-thread and short-lived.
-                }
-            }
-        }
-
-        return false;
-    }
-
-    private static IntPtr TrySetPerMonitorDpiAwareThread()
-    {
-        try
-        {
-            return NativeMethods.SetThreadDpiAwarenessContext(NativeMethods.DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2);
-        }
-        catch (Exception)
-        {
-            // SetThreadDpiAwarenessContext is unavailable before Windows 10 1607; proceed without it.
-            return IntPtr.Zero;
-        }
-    }
-
-    private static IEnumerable<IntPtr> EnumerateCandidateWindows()
-    {
-        // A GUI app under test owns its main window.
-        using (var current = Process.GetCurrentProcess())
-        {
-            yield return current.MainWindowHandle;
-        }
-
-        // The window you launched from (e.g. Windows Terminal), whose window is not owned by the
-        // test process. Captured at record time.
-        yield return NativeMethods.GetForegroundWindow();
-
-        // A classic console (conhost) window owned by the console host.
-        yield return NativeMethods.GetConsoleWindow();
-    }
-
-    private static void TryKill(Process process)
-    {
-        try
-        {
-            process.Kill();
-        }
-        catch (Exception)
-        {
-            // Best effort.
-        }
-    }
-
-    private static async Task<bool> WaitForExitAsync(Process process, TimeSpan timeout, CancellationToken cancellationToken)
-    {
-        // Process.WaitForExitAsync(CancellationToken) is available on modern .NET and provided by a
-        // polyfill on netstandard2.0, so a single token-based implementation works on every target.
-        // A linked CTS bounds the wait by the timeout; an already-cancelled token simply means we
-        // report whether the process happened to have exited (best-effort stop on a cancelled run).
-        using var cts = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
-        cts.CancelAfter(timeout);
-        try
-        {
-            await process.WaitForExitAsync(cts.Token).ConfigureAwait(false);
-            return true;
-        }
-        catch (OperationCanceledException)
-        {
-            return process.HasExited;
-        }
-    }
-
-    private static string? FindFfmpeg(string? configuredPath)
-    {
-        if (!RoslynString.IsNullOrEmpty(configuredPath))
-        {
-            return File.Exists(configuredPath) ? configuredPath : null;
-        }
-
-        string executable = RuntimeInformation.IsOSPlatform(OSPlatform.Windows) ? "ffmpeg.exe" : "ffmpeg";
-        string? pathVariable = Environment.GetEnvironmentVariable("PATH");
-        if (pathVariable is null)
-        {
-            return null;
-        }
-
-        foreach (string directory in pathVariable.Split(Path.PathSeparator))
-        {
-            if (RoslynString.IsNullOrWhiteSpace(directory))
-            {
-                continue;
-            }
-
-            try
-            {
-                string candidate = Path.Combine(directory.Trim(), executable);
-                if (File.Exists(candidate))
-                {
-                    return candidate;
-                }
-            }
-            catch (Exception)
-            {
-                // Ignore malformed PATH entries.
-            }
-        }
-
-        return null;
-    }
-
-    private static class NativeMethods
-    {
-        public const int SM_CXSCREEN = 0;
-        public const int SM_CYSCREEN = 1;
-
-        public static readonly IntPtr DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2 = new(-4);
-
-        [DllImport("kernel32.dll")]
-        public static extern IntPtr GetConsoleWindow();
-
-        [DllImport("user32.dll")]
-        public static extern IntPtr GetForegroundWindow();
-
-        [DllImport("user32.dll")]
-        public static extern int GetSystemMetrics(int nIndex);
-
-        [DllImport("user32.dll")]
-        public static extern IntPtr SetThreadDpiAwarenessContext(IntPtr dpiContext);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool IsWindowVisible(IntPtr hWnd);
-
-        [DllImport("user32.dll", SetLastError = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool GetWindowRect(IntPtr hWnd, out RECT lpRect);
-
-        [StructLayout(LayoutKind.Sequential)]
-        public struct RECT
-        {
-            public int Left;
-            public int Top;
-            public int Right;
-            public int Bottom;
-        }
     }
 }
